+++ src/router/routes/kompack/gudang/gudang.js
@@ -47,7 +47,7 @@ export default [
   {
     path: '/ajukan-inbound',
     name: 'ajukan-inbound',
-    component: () => import('@/views/pages/kompack/gudang/ajukan-inbound/ajukaninbound.vue'),
+    component: () => import('@/views/pages/kompack/gudang/ajukan-inbound/AjukanInbound.vue'),
     meta: {
       resource: 'Gudang',
       action: 'manage',
